Share embedded-server readiness probe across agent and proxy tests
Multiple tests followed the start(wait = false) + immediate-request
pattern, racing CIO's routing-pipeline install. Under SMALL/LARGE
harness load this surfaced as EOFException or empty response bodies
(e.g. ProxyHttpConfigTest StatusPages NotFound returning ""). Extract
AgentHttpServiceTest's HTTP-level probe into a shared
EmbeddedServer.startAndAwaitReady() extension and adopt it in the agent
header test and all 8 sites in ProxyHttpConfigTest.
