+++ src/test/kotlin/io/prometheus/agent/AgentHttpServiceTest.kt
@@ -30,7 +30,6 @@ import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldContain
 import io.ktor.http.HttpStatusCode
-import io.ktor.server.engine.EmbeddedServer
 import io.ktor.server.engine.embeddedServer
 import io.ktor.server.request.header
 import io.ktor.server.response.respondText
@@ -44,13 +43,10 @@ import io.prometheus.Agent
 import io.prometheus.common.ConfigVals
 import io.prometheus.grpc.registerPathResponse
 import io.prometheus.grpc.scrapeRequest
+import io.prometheus.common.startAndAwaitReady
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.delay
-import java.net.InetSocketAddress
-import java.net.Socket
 import kotlin.time.Duration.Companion.milliseconds
-import kotlin.time.Duration.Companion.seconds
-import kotlin.time.TimeSource.Monotonic
 import kotlin.time.measureTime
 import io.ktor.server.cio.CIO as ServerCIO
 
@@ -167,37 +163,6 @@ class AgentHttpServiceTest : StringSpec() {
     every { mockAgent.pathManager } returns pathManager
 
     return mockAgent
-  }
-
-  private suspend fun startServerAndGetPort(server: EmbeddedServer<*, *>): Int {
-    server.start(wait = false)
-    val port = server.engine.resolvedConnectors().first().port
-
-    // Poll until the server actually answers a complete HTTP exchange. A bare TCP-connect
-    // probe isn't enough: CIO's listening socket can accept and immediately close while
-    // the routing pipeline is still being installed, which makes the next real request
-    // race the install and come back with srValidResponse=false. Probing a path that
-    // won't match any registered route yields 404 once routing is live, which is the
-    // signal we want.
-    val deadline = Monotonic.markNow() + 5.seconds
-    while (Monotonic.markNow() < deadline) {
-      try {
-        Socket().use { sock ->
-          sock.connect(InetSocketAddress("localhost", port), 200)
-          sock.soTimeout = 200
-          sock.getOutputStream().apply {
-            write("GET /__readiness__ HTTP/1.0\r\nHost: localhost\r\n\r\n".toByteArray())
-            flush()
-          }
-          val reply = sock.getInputStream().readNBytes(12).decodeToString()
-          if (reply.startsWith("HTTP/")) return port
-        }
-      } catch (_: java.io.IOException) {
-        // not ready yet
-      }
-      delay(20.milliseconds)
-    }
-    error("Embedded server on port $port did not start serving HTTP within 5s")
   }
 
   init {
@@ -323,7 +288,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
 
         val mockAgent = createMockAgentWithPaths()
         val service = AgentHttpService(mockAgent)
@@ -387,7 +352,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
 
         val mockAgent = createMockAgentWithPaths()
         val service = AgentHttpService(mockAgent)
@@ -452,7 +417,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
 
         val mockAgent = createMockAgentWithPaths()
         val service = AgentHttpService(mockAgent)
@@ -494,7 +459,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
 
         val mockAgent = createMockAgentWithPaths()
         val service = AgentHttpService(mockAgent)
@@ -536,7 +501,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
 
         val mockAgent = createMockAgentWithPaths()
         // Set low gzip threshold so content gets zipped
@@ -579,7 +544,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
 
         val mockAgent = createMockAgentWithPaths()
         // minGzipSizeBytes is already 1_000_000 from createMockAgentWithPaths
@@ -623,7 +588,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
 
         val mockAgent = createMockAgentWithPaths()
         val service = AgentHttpService(mockAgent)
@@ -663,7 +628,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
 
         val mockAgent = createMockAgentWithPaths()
         val service = AgentHttpService(mockAgent)
@@ -704,7 +669,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
 
         val mockAgent = createMockAgentWithPaths()
         // Set a very short timeout
@@ -748,7 +713,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
 
         val mockAgent = createMockAgentWithPaths()
         val service = AgentHttpService(mockAgent)
@@ -791,7 +756,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
         val mockAgent = createMockAgentWithRetries(3)
         val service = AgentHttpService(mockAgent)
         mockAgent.pathManager.registerPath("metrics", "http://localhost:$port/metrics")
@@ -824,7 +789,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
         val mockAgent = createMockAgentWithRetries(3)
         val service = AgentHttpService(mockAgent)
         mockAgent.pathManager.registerPath("metrics", "http://localhost:$port/metrics")
@@ -857,7 +822,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
         val mockAgent = createMockAgentWithRetries(3)
         val service = AgentHttpService(mockAgent)
         mockAgent.pathManager.registerPath("metrics", "http://localhost:$port/metrics")
@@ -890,7 +855,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
         val mockAgent = createMockAgentWithRetries(2)
         val service = AgentHttpService(mockAgent)
         mockAgent.pathManager.registerPath("metrics", "http://localhost:$port/metrics")
@@ -925,7 +890,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
         val mockAgent = createMockAgentWithRetries(2)
         val service = AgentHttpService(mockAgent)
         mockAgent.pathManager.registerPath("metrics", "http://localhost:$port/metrics")
@@ -959,7 +924,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
         val mockAgent = createMockAgentWithRetries(3)
         val service = AgentHttpService(mockAgent)
         mockAgent.pathManager.registerPath("metrics", "http://localhost:$port/metrics")
@@ -1006,7 +971,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
 
         val mockAgent = createMockAgentWithPaths()
         // Set threshold between char count (200) and byte count (600)
@@ -1079,7 +1044,7 @@ class AgentHttpServiceTest : StringSpec() {
       }
 
       try {
-        val port = startServerAndGetPort(server)
+        val port = server.startAndAwaitReady()
 
         val mockAgent = createMockAgentWithRetries(10)
         // Set a short scrape timeout of 3 seconds
